refactor(phase10): extract typed-object/scalar helpers → codegen/typed_object_tokens.rs
Move 19 tightly cohesive helpers out of resource_struct_tokens.rs into a
new typed_object_tokens.rs sibling module:

  typed_object_validator_defs / typed_object_normalizer_defs
  field_needs_normalization
  collect_typed_object_normalizer_defs / collect_typed_object_validator_defs
  typed_object_validator_ident / typed_object_normalizer_ident
  string_transform_ops
  typed_object_validator_scalar_type_tokens
  typed_object_validator_list_item_type_tokens
  typed_object_validator_field_type_tokens
  typed_object_validator_field_check_tokens
  typed_object_validator_scalar_checks
  structured_scalar_type_tokens / structured_scalar_to_text_tokens
  json_bind_tokens
  list_field_from_text_tokens / structured_scalar_from_text_tokens
  field_supports_sort

resource_struct_tokens.rs: 1107 → 553 lines  (−554)
typed_object_tokens.rs:     606 lines (new)

resource_struct_tokens.rs imports the handful of these it still calls
directly (typed_object_validator_defs, typed_object_validator_ident,
list_field_from_text_tokens, structured_scalar_from_text_tokens,
field_supports_sort) via its `use super::{...}` block; all other callers
(resource_impl.rs, payload_tokens.rs, list_tokens.rs) already resolve
through codegen.rs's glob re-export chain.

Removed ToSnakeCase, ToUpperCamelCase, syn::Type from
resource_struct_tokens.rs — no longer needed after the extraction.

`cargo check --workspace` clean; 233/19 test baseline unchanged.
