+++ crates/rest_macro_core/src/compiler/codegen/resource_struct_tokens.rs
@@ -11,10 +11,9 @@
 //! Garde attribute helpers live in the sibling `garde_tokens` module and
 //! arrive here through the parent-module glob re-export.
 
-use heck::{ToSnakeCase, ToUpperCamelCase};
 use proc_macro2::{Literal, TokenStream};
 use quote::{format_ident, quote};
-use syn::{Path, Type};
+use syn::Path;
 
 use super::super::model::{ResourceSpec, WriteModelStyle};
 use super::{
@@ -23,17 +22,20 @@ use super::{
     // items via super::.
     ListBindKind,
     create_payload_field_is_optional, create_payload_field_ty, create_payload_fields,
-    enum_values_as_message,
     // garde helpers (from garde_tokens via parent glob re-export)
     garde_field_attr_tokens,
     garde_validation_helper_defs,
-    prefixed_garde_validation_error_helper,
     list_bind_ident, list_bind_kinds,
     list_cursor_payload_ident, list_cursor_value_ident,
     list_filter_field_ty, list_plan_ident,
     list_query_ident, list_response_ident,
     list_sort_field_ident, list_sort_order_ident,
     resource_action_input_struct_tokens,
+    // typed_object helpers (from typed_object_tokens via parent glob re-export)
+    field_supports_sort,
+    list_field_from_text_tokens,
+    structured_scalar_from_text_tokens,
+    typed_object_validator_defs, typed_object_validator_ident,
     update_payload_fields,
 };
 use crate::authorization::AuthorizationContract;
@@ -184,562 +186,6 @@ pub(super) fn serde_rename_attr(api_name: &str, storage_name: &str) -> TokenStre
     }
 }
 
-pub(super) fn typed_object_validator_defs(resource: &ResourceSpec, runtime_crate: &Path) -> Vec<TokenStream> {
-    resource
-        .api_fields()
-        .flat_map(|field| {
-            let path = vec![field.name()];
-            collect_typed_object_validator_defs(resource, field, &path, runtime_crate)
-        })
-        .collect()
-}
-
-pub(super) fn typed_object_normalizer_defs(resource: &ResourceSpec, runtime_crate: &Path) -> Vec<TokenStream> {
-    resource
-        .api_fields()
-        .flat_map(|field| {
-            let path = vec![field.name()];
-            collect_typed_object_normalizer_defs(resource, field, &path, runtime_crate)
-        })
-        .collect()
-}
-
-pub(super) fn field_needs_normalization(field: &super::super::model::FieldSpec) -> bool {
-    !field.transforms().is_empty()
-        || field
-            .object_fields
-            .as_deref()
-            .map(|nested_fields| nested_fields.iter().any(field_needs_normalization))
-            .unwrap_or(false)
-}
-
-pub(super) fn collect_typed_object_normalizer_defs(
-    resource: &ResourceSpec,
-    field: &super::super::model::FieldSpec,
-    path: &[String],
-    runtime_crate: &Path,
-) -> Vec<TokenStream> {
-    let Some(nested_fields) = field.object_fields.as_deref() else {
-        return Vec::new();
-    };
-
-    let mut definitions = Vec::new();
-    for nested_field in nested_fields {
-        let mut nested_path = path.to_vec();
-        nested_path.push(nested_field.name());
-        definitions.extend(collect_typed_object_normalizer_defs(
-            resource,
-            nested_field,
-            &nested_path,
-            runtime_crate,
-        ));
-    }
-
-    let helper_ident = typed_object_normalizer_ident(resource, path);
-    let statements = nested_fields
-        .iter()
-        .filter_map(|nested_field| {
-            let field_name = Literal::string(nested_field.api_name());
-            if nested_field.object_fields.is_some() {
-                if !field_needs_normalization(nested_field) {
-                    return None;
-                }
-                let mut nested_path = path.to_vec();
-                nested_path.push(nested_field.name());
-                let nested_helper_ident = typed_object_normalizer_ident(resource, &nested_path);
-                Some(quote! {
-                    if let Some(value) = object.get_mut(#field_name) {
-                        Self::#nested_helper_ident(value);
-                    }
-                })
-            } else if !nested_field.transforms().is_empty() {
-                let transform_ops = string_transform_ops(nested_field.transforms());
-                Some(quote! {
-                    if let Some(value) = object.get_mut(#field_name) {
-                        if let #runtime_crate::serde_json::Value::String(value) = value {
-                            #(#transform_ops)*
-                        }
-                    }
-                })
-            } else {
-                None
-            }
-        })
-        .collect::<Vec<_>>();
-
-    if statements.is_empty() {
-        return definitions;
-    }
-
-    definitions.push(quote! {
-        fn #helper_ident(value: &mut #runtime_crate::serde_json::Value) {
-            let Some(object) = value.as_object_mut() else {
-                return;
-            };
-            #(#statements)*
-        }
-    });
-
-    definitions
-}
-
-pub(super) fn collect_typed_object_validator_defs(
-    resource: &ResourceSpec,
-    field: &super::super::model::FieldSpec,
-    path: &[String],
-    runtime_crate: &Path,
-) -> Vec<TokenStream> {
-    let Some(nested_fields) = field.object_fields.as_deref() else {
-        return Vec::new();
-    };
-
-    let mut definitions = Vec::new();
-    for nested_field in nested_fields {
-        let mut nested_path = path.to_vec();
-        nested_path.push(nested_field.name());
-        definitions.extend(collect_typed_object_validator_defs(
-            resource,
-            nested_field,
-            &nested_path,
-            runtime_crate,
-        ));
-    }
-
-    let struct_ident = typed_object_validator_ident(resource, path);
-    let field_defs = nested_fields.iter().map(|nested_field| {
-        let ident = &nested_field.ident;
-        let mut nested_path = path.to_vec();
-        nested_path.push(nested_field.name());
-        let ty = typed_object_validator_field_type_tokens(
-            resource,
-            nested_field,
-            &nested_path,
-            runtime_crate,
-        );
-        let rename_attr = serde_rename_attr(nested_field.api_name(), &nested_field.name());
-        let garde_attr = garde_field_attr_tokens(
-            nested_field,
-            super::super::model::is_optional_type(&nested_field.ty),
-            true,
-        );
-        if super::super::model::is_optional_type(&nested_field.ty) {
-            quote! {
-                #rename_attr
-                #garde_attr
-                #[serde(default)]
-                pub #ident: #ty,
-            }
-        } else {
-            quote! {
-                #rename_attr
-                #garde_attr
-                pub #ident: #ty,
-            }
-        }
-    });
-    let validations = nested_fields
-        .iter()
-        .map(typed_object_validator_field_check_tokens);
-    let prefixed_error_helper = prefixed_garde_validation_error_helper(resource);
-
-    definitions.push(quote! {
-        #[allow(dead_code)]
-        #[derive(
-            Debug,
-            Clone,
-            #runtime_crate::serde::Deserialize,
-            #runtime_crate::garde::Validate
-        )]
-        #[garde(allow_unvalidated)]
-        #[serde(deny_unknown_fields)]
-        struct #struct_ident {
-            #(#field_defs)*
-        }
-
-        impl #struct_ident {
-            fn validate(&self, base_path: &str) -> Result<(), (String, String)> {
-                if let Err(report) = #runtime_crate::garde::Validate::validate(self) {
-                    return Err(#prefixed_error_helper(base_path, report));
-                }
-                #(#validations)*
-                Ok(())
-            }
-        }
-    });
-
-    definitions
-}
-
-pub(super) fn typed_object_validator_ident(resource: &ResourceSpec, path: &[String]) -> syn::Ident {
-    let suffix = path
-        .iter()
-        .map(|segment| segment.to_upper_camel_case())
-        .collect::<String>();
-    format_ident!("{}{}ObjectSchema", resource.struct_ident, suffix)
-}
-
-pub(super) fn typed_object_normalizer_ident(resource: &ResourceSpec, path: &[String]) -> syn::Ident {
-    let suffix = path
-        .iter()
-        .map(|segment| segment.to_upper_camel_case())
-        .collect::<String>();
-    format_ident!(
-        "normalize_{}_{}_object",
-        resource.struct_ident.to_string().to_snake_case(),
-        suffix.to_snake_case()
-    )
-}
-
-pub(super) fn string_transform_ops(transforms: &[super::super::model::FieldTransform]) -> Vec<TokenStream> {
-    transforms
-        .iter()
-        .map(|transform| match transform {
-            super::super::model::FieldTransform::Trim => quote! {
-                *value = value.trim().to_owned();
-            },
-            super::super::model::FieldTransform::Lowercase => quote! {
-                *value = value.to_lowercase();
-            },
-            super::super::model::FieldTransform::CollapseWhitespace => quote! {
-                *value = value.split_whitespace().collect::<Vec<_>>().join(" ");
-            },
-            super::super::model::FieldTransform::Slugify => quote! {
-                {
-                    let mut slug = String::new();
-                    let mut pending_dash = false;
-                    for ch in value.chars() {
-                        if ch.is_alphanumeric() {
-                            if pending_dash && !slug.is_empty() {
-                                slug.push('-');
-                            }
-                            pending_dash = false;
-                            for lower in ch.to_lowercase() {
-                                slug.push(lower);
-                            }
-                        } else if !slug.is_empty() {
-                            pending_dash = true;
-                        }
-                    }
-                    *value = slug;
-                }
-            },
-        })
-        .collect()
-}
-
-pub(super) fn typed_object_validator_scalar_type_tokens(ty: &Type, runtime_crate: &Path) -> TokenStream {
-    let base_ty = super::super::model::base_type(ty);
-    if let Some(kind) = super::super::model::structured_scalar_kind(&base_ty) {
-        return match kind {
-            super::super::model::StructuredScalarKind::Json => {
-                quote!(#runtime_crate::serde_json::Value)
-            }
-            super::super::model::StructuredScalarKind::JsonObject => {
-                quote!(#runtime_crate::serde_json::Map<String, #runtime_crate::serde_json::Value>)
-            }
-            super::super::model::StructuredScalarKind::JsonArray => {
-                quote!(Vec<#runtime_crate::serde_json::Value>)
-            }
-            _ => structured_scalar_type_tokens(kind, runtime_crate),
-        };
-    }
-
-    if super::super::model::is_bool_type(&base_ty) {
-        quote!(bool)
-    } else {
-        quote!(#base_ty)
-    }
-}
-
-pub(super) fn typed_object_validator_list_item_type_tokens(ty: &Type, runtime_crate: &Path) -> TokenStream {
-    if let Some(kind) = super::super::model::structured_scalar_kind(ty) {
-        return match kind {
-            super::super::model::StructuredScalarKind::Json => {
-                quote!(#runtime_crate::serde_json::Value)
-            }
-            super::super::model::StructuredScalarKind::JsonObject => {
-                quote!(#runtime_crate::serde_json::Map<String, #runtime_crate::serde_json::Value>)
-            }
-            super::super::model::StructuredScalarKind::JsonArray => {
-                quote!(Vec<#runtime_crate::serde_json::Value>)
-            }
-            _ => structured_scalar_type_tokens(kind, runtime_crate),
-        };
-    }
-
-    if super::super::model::is_bool_type(ty) {
-        quote!(bool)
-    } else {
-        quote!(#ty)
-    }
-}
-
-pub(super) fn typed_object_validator_field_type_tokens(
-    resource: &ResourceSpec,
-    field: &super::super::model::FieldSpec,
-    path: &[String],
-    runtime_crate: &Path,
-) -> TokenStream {
-    let inner_ty = if field.object_fields.is_some() {
-        let ident = typed_object_validator_ident(resource, path);
-        quote!(#ident)
-    } else if let Some(item_ty) = field.list_item_ty.as_ref() {
-        let item_ty = typed_object_validator_list_item_type_tokens(item_ty, runtime_crate);
-        quote!(Vec<#item_ty>)
-    } else {
-        typed_object_validator_scalar_type_tokens(&field.ty, runtime_crate)
-    };
-
-    if super::super::model::is_optional_type(&field.ty) {
-        quote!(Option<#inner_ty>)
-    } else {
-        inner_ty
-    }
-}
-
-pub(super) fn typed_object_validator_field_check_tokens(field: &super::super::model::FieldSpec) -> TokenStream {
-    let ident = &field.ident;
-    let field_name = field.api_name().to_owned();
-    let field_name_lit = Literal::string(&field_name);
-
-    if field.object_fields.is_some() {
-        return if super::super::model::is_optional_type(&field.ty) {
-            quote! {
-                if let Some(value) = &self.#ident {
-                    let nested_path = format!("{}.{}", base_path, #field_name_lit);
-                    value.validate(nested_path.as_str())?;
-                }
-            }
-        } else {
-            quote! {
-                let nested_path = format!("{}.{}", base_path, #field_name_lit);
-                self.#ident.validate(nested_path.as_str())?;
-            }
-        };
-    }
-
-    if field.enum_values().is_none() {
-        return quote! {};
-    }
-
-    let checks = typed_object_validator_scalar_checks(field);
-    if checks.is_empty() {
-        return quote! {};
-    }
-
-    if super::super::model::is_optional_type(&field.ty) {
-        quote! {
-            if let Some(value) = &self.#ident {
-                let field_path = format!("{}.{}", base_path, #field_name_lit);
-                #(#checks)*
-            }
-        }
-    } else {
-        quote! {
-            let value = &self.#ident;
-            let field_path = format!("{}.{}", base_path, #field_name_lit);
-            #(#checks)*
-        }
-    }
-}
-
-pub(super) fn typed_object_validator_scalar_checks(field: &super::super::model::FieldSpec) -> Vec<TokenStream> {
-    let mut checks = Vec::new();
-
-    if let Some(enum_values) = field.enum_values() {
-        let enum_values = enum_values
-            .iter()
-            .map(|value| Literal::string(value.as_str()))
-            .collect::<Vec<_>>();
-        let enum_values_message = Literal::string(&enum_values_as_message(field));
-        checks.push(quote! {
-            if ![#(#enum_values),*].contains(&value.as_str()) {
-                return Err((
-                    field_path.clone(),
-                    format!("Field `{}` must be one of: {}", field_path, #enum_values_message),
-                ));
-            }
-        });
-    }
-
-    checks
-}
-
-pub(super) fn structured_scalar_type_tokens(
-    kind: super::super::model::StructuredScalarKind,
-    runtime_crate: &Path,
-) -> TokenStream {
-    match kind {
-        super::super::model::StructuredScalarKind::DateTime => {
-            quote!(#runtime_crate::chrono::DateTime<#runtime_crate::chrono::Utc>)
-        }
-        super::super::model::StructuredScalarKind::Date => quote!(#runtime_crate::chrono::NaiveDate),
-        super::super::model::StructuredScalarKind::Time => quote!(#runtime_crate::chrono::NaiveTime),
-        super::super::model::StructuredScalarKind::Uuid => quote!(#runtime_crate::uuid::Uuid),
-        super::super::model::StructuredScalarKind::Decimal => {
-            quote!(#runtime_crate::rust_decimal::Decimal)
-        }
-        super::super::model::StructuredScalarKind::Json
-        | super::super::model::StructuredScalarKind::JsonObject
-        | super::super::model::StructuredScalarKind::JsonArray => {
-            quote!(#runtime_crate::serde_json::Value)
-        }
-    }
-}
-
-pub(super) fn structured_scalar_to_text_tokens(
-    ty: &syn::Type,
-    value: TokenStream,
-    runtime_crate: &Path,
-) -> Option<TokenStream> {
-    match super::super::model::structured_scalar_kind(ty)? {
-        super::super::model::StructuredScalarKind::DateTime => Some(quote! {
-            #value.to_rfc3339_opts(#runtime_crate::chrono::SecondsFormat::Micros, false)
-        }),
-        super::super::model::StructuredScalarKind::Date => {
-            Some(quote!(#value.format("%Y-%m-%d").to_string()))
-        }
-        super::super::model::StructuredScalarKind::Time => {
-            Some(quote!(#value.format("%H:%M:%S.%6f").to_string()))
-        }
-        super::super::model::StructuredScalarKind::Uuid => {
-            Some(quote!(#value.as_hyphenated().to_string()))
-        }
-        super::super::model::StructuredScalarKind::Decimal => Some(quote!(#value.normalize().to_string())),
-        super::super::model::StructuredScalarKind::Json
-        | super::super::model::StructuredScalarKind::JsonObject
-        | super::super::model::StructuredScalarKind::JsonArray => Some(quote!(
-            #runtime_crate::serde_json::to_string(#value).expect("JSON fields should serialize")
-        )),
-    }
-}
-
-pub(super) fn json_bind_tokens(field: &super::super::model::FieldSpec, runtime_crate: &Path) -> Option<TokenStream> {
-    if field.list_item_ty.is_some() {
-        let ident = &field.ident;
-        if super::super::model::is_optional_type(&field.ty) {
-            return Some(quote! {
-                item.#ident.as_ref().map(|value| {
-                    #runtime_crate::serde_json::to_string(value)
-                        .expect("list fields should serialize")
-                })
-            });
-        }
-        return Some(quote! {
-            #runtime_crate::serde_json::to_string(&item.#ident)
-                .expect("list fields should serialize")
-        });
-    }
-
-    match super::super::model::structured_scalar_kind(&field.ty) {
-        Some(
-            super::super::model::StructuredScalarKind::Json
-            | super::super::model::StructuredScalarKind::JsonObject
-            | super::super::model::StructuredScalarKind::JsonArray,
-        ) => {
-            let ident = &field.ident;
-            if super::super::model::is_optional_type(&field.ty) {
-                Some(quote! {
-                    item.#ident.as_ref().map(|value| {
-                        #runtime_crate::serde_json::to_string(value)
-                            .expect("JSON fields should serialize")
-                    })
-                })
-            } else {
-                Some(quote! {
-                    #runtime_crate::serde_json::to_string(&item.#ident)
-                        .expect("JSON fields should serialize")
-                })
-            }
-        }
-        _ => None,
-    }
-}
-
-pub(super) fn list_field_from_text_tokens(
-    base_ty: &Type,
-    value_ident: TokenStream,
-    field_name_lit: &Literal,
-    runtime_crate: &Path,
-) -> TokenStream {
-    quote! {
-        #runtime_crate::serde_json::from_str::<#base_ty>(&#value_ident).map_err(
-            |error| #runtime_crate::sqlx::Error::ColumnDecode {
-                index: #field_name_lit.to_owned(),
-                source: Box::new(error),
-            }
-        )?
-    }
-}
-
-pub(super) fn structured_scalar_from_text_tokens(
-    kind: super::super::model::StructuredScalarKind,
-    base_ty: &Type,
-    value_ident: TokenStream,
-    field_name_lit: &Literal,
-    runtime_crate: &Path,
-) -> TokenStream {
-    match kind {
-        super::super::model::StructuredScalarKind::DateTime
-        | super::super::model::StructuredScalarKind::Date
-        | super::super::model::StructuredScalarKind::Time
-        | super::super::model::StructuredScalarKind::Uuid
-        | super::super::model::StructuredScalarKind::Decimal => quote! {
-            #value_ident.parse::<#base_ty>().map_err(|error| #runtime_crate::sqlx::Error::ColumnDecode {
-                index: #field_name_lit.to_owned(),
-                source: Box::new(error),
-            })?
-        },
-        super::super::model::StructuredScalarKind::Json => quote! {
-            #runtime_crate::serde_json::from_str::<#base_ty>(&#value_ident).map_err(
-                |error| #runtime_crate::sqlx::Error::ColumnDecode {
-                    index: #field_name_lit.to_owned(),
-                    source: Box::new(error),
-                }
-            )?
-        },
-        super::super::model::StructuredScalarKind::JsonObject => quote! {{
-            let parsed = #runtime_crate::serde_json::from_str::<#base_ty>(&#value_ident).map_err(
-                |error| #runtime_crate::sqlx::Error::ColumnDecode {
-                    index: #field_name_lit.to_owned(),
-                    source: Box::new(error),
-                }
-            )?;
-            if !matches!(parsed, #runtime_crate::serde_json::Value::Object(_)) {
-                return Err(#runtime_crate::sqlx::Error::ColumnDecode {
-                    index: #field_name_lit.to_owned(),
-                    source: Box::new(::std::io::Error::new(
-                        ::std::io::ErrorKind::InvalidData,
-                        "expected JSON object",
-                    )),
-                });
-            }
-            parsed
-        }},
-        super::super::model::StructuredScalarKind::JsonArray => quote! {{
-            let parsed = #runtime_crate::serde_json::from_str::<#base_ty>(&#value_ident).map_err(
-                |error| #runtime_crate::sqlx::Error::ColumnDecode {
-                    index: #field_name_lit.to_owned(),
-                    source: Box::new(error),
-                }
-            )?;
-            if !matches!(parsed, #runtime_crate::serde_json::Value::Array(_)) {
-                return Err(#runtime_crate::sqlx::Error::ColumnDecode {
-                    index: #field_name_lit.to_owned(),
-                    source: Box::new(::std::io::Error::new(
-                        ::std::io::ErrorKind::InvalidData,
-                        "expected JSON array",
-                    )),
-                });
-            }
-            parsed
-        }},
-    }
-}
-
-pub(super) fn field_supports_sort(field: &super::super::model::FieldSpec) -> bool {
-    super::super::model::supports_field_sort(field)
-}
 
 pub(super) fn generated_from_row_tokens(resource: &ResourceSpec, runtime_crate: &Path) -> TokenStream {
     let struct_ident = &resource.struct_ident;
